Fix ScalaJS
